pyproject.toml
[project]
name = "vibe_yaml"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
license = "wtfpl"
license-files = ["LICEN[CS]E*"]
maintainers = [
	{name = "Mia Rose Winter", email = "mia.winter@winter-software.com"}
]
classifiers = [
	"Programming Language :: Python :: 3"
]
dependencies = [
    "annotated-doc==0.0.4",
    "anyio==4.12.1",
    "certifi==2026.2.25",
    "click==8.3.1",
    "filelock==3.25.2",
    "fsspec==2026.2.0",
    "h11==0.16.0",
    "hf-xet==1.4.2",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.7.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "networkx==3.6.1",
    "numpy==2.4.3",
    "packaging==26.0",
    "pygments==2.19.2",
    "pyyaml==6.0.3",
    "regex==2026.2.28",
    "rich==14.3.3",
    "safetensors==0.7.0",
    "setuptools==70.2.0",
    "shellingham==1.5.4",
    "sympy==1.14.0",
    "tokenizers==0.22.2",
    "torch==2.10.0+cpu",
    "tqdm==4.67.3",
    "transformers==5.3.0",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
]

[project.urls]
Homepage = "https://github.com/miawinter98/vibe-yaml"
Issues = "https://github.com/miawinter98/vibe-yaml/issues"

